Create the option file, and add the option class
