+++ main.rb
@@ -1,16 +1,6 @@
 require_relative './app'
+require_relative './options'
 
-def print_options
-  puts ''
-  puts 'Please choose and option by entering a number:'
-  puts '1 - List all books'
-  puts '2 - List all people'
-  puts '3 - Create a person'
-  puts '4 - Create a book'
-  puts '5 - Create a rental'
-  puts '6 - List all the rentals for a given person id'
-  puts '7 - Exit'
-end
 
 def call_actions(option, app)
   case option
