+++ options.rb
@@ -1,0 +1,13 @@
+class Option
+  def print_options
+    puts ''
+    puts 'Please choose and option by entering a number:'
+    puts '1 - List all books'
+    puts '2 - List all people'
+    puts '3 - Create a person'
+    puts '4 - Create a book'
+    puts '5 - Create a rental'
+    puts '6 - List all the rentals for a given person id'
+    puts '7 - Exit'
+  end
+end
